# packages/contracts/rest/auth.openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service (Internal)
  version: 0.2.0
servers:
  - url: http://auth_service:3000
paths:
  /internal/auth:
    get:
      summary: Auth service root
      responses:
        "200": { description: OK }

  /internal/auth/health:
    get:
      summary: Auth service health
      responses:
        "200": { description: OK }

  /internal/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "201": { description: Created }

  /internal/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/google/exchange:
    post:
      summary: Exchange Google auth
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/refresh:
    post:
      summary: Refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/logout:
    post:
      summary: Logout session(s)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/verify:
    get:
      summary: Verify access token
      responses:
        "200": { description: OK }

  /internal/auth/users/search:
    get:
      summary: Search active users by username, email, or id
      parameters:
        - in: query
          name: query
          required: false
          schema: { type: string }
        - in: query
          name: cursor
          required: false
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200": { description: OK }

  /internal/auth/password/set:
    post:
      summary: Set password
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/audit:
    get:
      summary: List auth audit logs
      responses:
        "200": { description: OK }

  /internal/auth/users:
    get:
      summary: Search users
      responses:
        "200": { description: OK }

  /internal/auth/users/{userId}:
    get:
      summary: Get user by id
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }

  /internal/auth/users/{userId}/disable:
    post:
      summary: Disable user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/users/{userId}/enable:
    post:
      summary: Enable user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/users/{userId}/role:
    post:
      summary: Set user roles
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }

  /internal/auth/users/{userId}/sessions/revoke:
    post:
      summary: Revoke user sessions
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }